requirements.txt
pandas==2.2.0
requests==2.31.0
selenium==4.19.0
webdriver-manager==4.0.1
stravalib==1.5.0
python-dotenv==1.0.1
psutil==5.9.8
beautifulsoup4==4.12.3
tcxreader
sqlalchemy==2.0.23
pytest==7.4.4
black==24.4.2
isort==5.13.2
mypy==1.8.0
tqdm==4.66.4
